requirements.txt
customtkinter
firebase-admin
reportlab
tkcalendar
pillow
